ogg rate, 32bit rate
git-svn-id: svn://elaine/murooma/trunk@288 d8a302eb-03bc-478d-80e4-98257eca68ef

+++ message/sampleFormat.cpp
@@ -16,7 +16,7 @@ SampleFormat::SampleFormat(const std::string& format) : BaseMessage(message_type
 }
 
 
-SampleFormat::SampleFormat(uint16_t sampleRate, uint16_t bitsPerSample, uint16_t channelCount) : BaseMessage(message_type::sampleformat)
+SampleFormat::SampleFormat(uint32_t sampleRate, uint16_t bitsPerSample, uint16_t channelCount) : BaseMessage(message_type::sampleformat)
 {
 	setFormat(sampleRate, bitsPerSample, channelCount);
 }
@@ -28,13 +28,13 @@ void SampleFormat::setFormat(const std::string& format)
 	boost::split(strs, format, boost::is_any_of(":"));
 	if (strs.size() == 3)
 		setFormat(
-		    boost::lexical_cast<uint16_t>(strs[0]),
+		    boost::lexical_cast<uint32_t>(strs[0]),
 		    boost::lexical_cast<uint16_t>(strs[1]),
 		    boost::lexical_cast<uint16_t>(strs[2]));
 }
 
 
-void SampleFormat::setFormat(uint16_t rate, uint16_t bits, uint16_t channels)
+void SampleFormat::setFormat(uint32_t rate, uint16_t bits, uint16_t channels)
 {
 	this->rate = rate;
 	this->bits = bits;
